kamael gameserver merge with current il trunk
git-svn-id: http://svn.l2jfree.com/l2jfree@2118 6ff00ba9-25cd-4355-987e-099c62bc2628

+++ branches/L2_GameServer_CT1/src/main/java/net/sf/l2j/gameserver/characters/model/recommendation/CharRecommendation.java
@@ -61,8 +61,8 @@ public class CharRecommendation
     {
         final int PRIME = 31;
         int result = 1;
-        result = PRIME * result + (int) (charId ^ (charId >>> 32));
-        result = PRIME * result + (int) (targetId ^ (targetId >>> 32));
+        result = PRIME * result + (charId ^ (charId >>> 32));
+        result = PRIME * result + (targetId ^ (targetId >>> 32));
         return result;
     }
     /* (non-Javadoc)
